% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R, R/file_system_async.R
\name{delete}
\alias{delete}
\alias{s3_file_delete}
\alias{s3_dir_delete}
\alias{s3_file_delete_async}
\alias{s3_dir_delete_async}
\title{Delete files and directories}
\usage{
s3_file_delete(path, ...)

s3_dir_delete(path)

s3_file_delete_async(path, ...)

s3_dir_delete_async(path)
}
\arguments{
\item{path}{(character): A character vector of paths or s3 uris.}

\item{...}{parameters to be passed to \code{\link[paws.storage]{s3_delete_objects}}}
}
\description{
\code{s3_file_delete} delete files in AWS S3

\code{s3_dir_delete} delete directories in AWS S3 recursively.
}
