% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system_async.R
\name{delete_async}
\alias{delete_async}
\alias{s3_file_delete_async}
\alias{s3_dir_delete_async}
\title{Delete files and directories}
\usage{
s3_file_delete_async(path, ...)

s3_dir_delete_async(path)
}
\arguments{
\item{path}{(character): A character vector of paths or s3 uris.}

\item{...}{parameters to be passed to \code{\link[paws.storage]{s3_delete_objects}}}
}
\value{
return \code{\link[future]{future}} object of \code{\link[=s3_file_delete]{s3_file_delete()}} \code{\link[=s3_dir_delete]{s3_dir_delete()}}
}
\description{
\code{s3_file_delete} delete files in AWS S3

\code{s3_dir_delete} delete directories in AWS S3 recursively.
}
\seealso{
\code{\link[future]{future}} \code{\link[=s3_file_delete]{s3_file_delete()}} \code{\link[=s3_dir_delete]{s3_dir_delete()}}
}
