% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{exists}
\alias{exists}
\alias{s3_file_exists}
\alias{s3_dir_exists}
\title{Download files and directories}
\usage{
s3_file_exists(path)

s3_dir_exists(path = ".")
}
\arguments{
\item{path}{(character) s3 path to check}
}
\description{
\code{s3_file_exists} check if file exists in AWS S3

\code{s3_dir_exists} check if path is a directory in AWS S3
}
