% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{file_type}
\alias{file_type}
\alias{s3_is_file}
\alias{s3_is_dir}
\alias{s3_is_bucket}
\alias{s3_is_file_empty}
\title{Functions to test for file types}
\usage{
s3_is_file(path)

s3_is_dir(path)

s3_is_bucket(path, ...)

s3_is_file_empty(path)
}
\arguments{
\item{path}{(character): A character vector of paths or uris}

\item{...}{parameters to be passed to \code{\link[paws.storage]{s3_list_objects_v2}}}
}
\description{
Test for file types
}
