% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{path}
\alias{path}
\alias{s3_path}
\title{Construct path for file or directory}
\usage{
s3_path(..., ext = "")
}
\arguments{
\item{...}{(character): Character vectors}

\item{ext}{(character): An optional extension to append to the generated path}
}
\description{
Constructs a s3 uri path
}
