% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{s3_bucket_delete}
\alias{s3_bucket_delete}
\title{Delete bucket}
\usage{
s3_bucket_delete(path)
}
\arguments{
\item{path}{(character): A character vector of path or s3 uri.}
}
\description{
Delete \verb{AWS S3} bucket including all objects in the bucket itself.
}
