% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{s3_dir_ls_url}
\alias{s3_dir_ls_url}
\title{Generate presigned url to list S3 directories}
\usage{
s3_dir_ls_url(path, expiration = 3600L, recurse = FALSE, ...)
}
\arguments{
\item{path}{(character): A character vector of paths or uris}

\item{expiration}{(numeric): The number of seconds the presigned url is
valid for. By default it expires in an hour (3600 seconds)}

\item{recurse}{(logical): Returns all AWS S3 objects in lower sub directories}

\item{...}{parameters passed to \code{\link[paws.storage]{s3_list_objects_v2}}}
}
\value{
return character of urls
}
\description{
Generate presigned url to list S3 directories
}
