% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{s3_dir_tree}
\alias{s3_dir_tree}
\title{Print contents of directories in a tree-like format}
\usage{
s3_dir_tree(path, recurse = TRUE, ...)
}
\arguments{
\item{path}{(character): path A path to print the tree from}

\item{recurse}{(logical): Returns all AWS S3 objects in lower sub directories}

\item{...}{Additional arguments passed to \link{s3_dir_ls}.}
}
\value{
character vector of s3 uri paths
}
\description{
Print contents of directories in a tree-like format
}
