% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{s3_file_system}
\alias{s3_file_system}
\title{Access AWS S3 as if it were a file system.}
\usage{
s3_file_system(
  aws_access_key_id = NULL,
  aws_secret_access_key = NULL,
  aws_session_token = NULL,
  region_name = NULL,
  profile_name = NULL,
  endpoint = NULL,
  disable_ssl = FALSE,
  refresh = FALSE,
  ...
)
}
\arguments{
\item{aws_access_key_id}{(character): AWS access key ID}

\item{aws_secret_access_key}{(character): AWS secret access key}

\item{aws_session_token}{(character): AWS temporary session token}

\item{region_name}{(character): Default region when creating new connections}

\item{profile_name}{(character): The name of a profile to use. If not given,
then the default profile is used.}

\item{endpoint}{(character): The complete URL to use for the constructed client.}

\item{disable_ssl}{(logical): Whether or not to use SSL. By default, SSL is used.}

\item{refresh}{(logical): Refresh cached S3FileSystem class}

\item{...}{Other parameters within \code{paws} client.}
}
\description{
This creates a file system "like" API based off \code{fs}
(e.g. dir_ls, file_copy, etc.) for AWS S3 storage.
}
