% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{s3_file_temp}
\alias{s3_file_temp}
\title{Create name for temporary files}
\usage{
s3_file_temp(pattern = "file", tmp_dir = "", ext = "")
}
\arguments{
\item{pattern}{(character): A character vector with the non-random portion of the name.}

\item{tmp_dir}{(character): The directory the file will be created in. By default
the cached s3 bucket will be applied otherwise \code{""} will be used.}

\item{ext}{(character): A character vector of one or more paths.}
}
\value{
character vector of s3 uri paths
}
\description{
return the name which can be used as a temporary file
}
\examples{
\dontrun{
# Require AWS S3 credentials

s3_file_temp(tmp_dir = "MyBucket")
}
}
