% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{s3_file_url}
\alias{s3_file_url}
\title{Generate presigned url for S3 object}
\usage{
s3_file_url(path, expiration = 3600L, ...)
}
\arguments{
\item{path}{(character): A character vector of paths or uris}

\item{expiration}{(numeric): The number of seconds the presigned url is
valid for. By default it expires in an hour (3600 seconds)}

\item{...}{parameters to be passed to \code{params} parameter of
\code{\link[paws.storage]{s3_generate_presigned_url}}}
}
\value{
return character of urls
}
\description{
Generate presigned url for S3 object
}
