% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{s3_file_version_info}
\alias{s3_file_version_info}
\title{Query file version metadata}
\usage{
s3_file_version_info(path, ...)
}
\arguments{
\item{path}{(character): A character vector of paths or uris}

\item{...}{parameters to be passed to \code{\link[paws.storage]{s3_list_object_versions}}}
}
\description{
Get file versions
}
