% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{s3_path_split}
\alias{s3_path_split}
\title{Split s3 path and uri}
\usage{
s3_path_split(path)
}
\arguments{
\item{path}{(character): A character vector of one or more paths or s3 uri}
}
\description{
Split s3 uri path to core components bucket, key and version id
}
