% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{tag}
\alias{tag}
\alias{s3_file_tag_delete}
\alias{s3_file_tag_info}
\alias{s3_file_tag_update}
\title{Modifying file tags}
\usage{
s3_file_tag_delete(path)

s3_file_tag_info(path)

s3_file_tag_update(path, tags, overwrite = FALSE)
}
\arguments{
\item{path}{(character): A character vector of paths or s3 uri}

\item{tags}{(list): Tags to be applied}

\item{overwrite}{(logical): To overwrite tagging or to modify inplace. Default will
modify inplace.}
}
\description{
\code{s3_file_tag_delete} delete file tags

\code{s3_file_tag_info} get file tags

\code{s3_file_tag_info}
}
