% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R, R/file_system_async.R
\name{touch}
\alias{touch}
\alias{s3_file_touch}
\alias{s3_file_touch_async}
\title{Change file modification time}
\usage{
s3_file_touch(path, ...)

s3_file_touch_async(path, ...)
}
\arguments{
\item{path}{(character): A character vector of paths or s3 uri}

\item{...}{parameters to be passed to \code{\link[paws.storage]{s3_copy_object}}}
}
\description{
Similar to \code{fs::file_touch} this does not create the file if
it does not exist. Use \code{s3fs$file_create()} to do this if needed.
}
\note{
This method will only update the modification time of the AWS S3 object.
}
