% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{path}
\alias{path}
\alias{s3_path}
\title{Construct path for file or directory}
\usage{
s3_path(..., ext = "")
}
\arguments{
\item{...}{(character): Character vectors}

\item{ext}{(character): An optional extension to append to the generated path}
}
\value{
character vector of s3 uri paths
}
\description{
Constructs a s3 uri path
}
\examples{
\dontrun{
# Require AWS S3 credentials

s3_path("my_bucket1", "my_bucket2")
}
}
