% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{s3fs-package}
\alias{s3fs}
\alias{s3fs-package}
\title{s3fs: Amazon Web Service 'S3' File System}
\description{
Access Amazon Web Service 'S3' as if it were a file system. 'API' is based upon 'fs' package.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

}
