% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{s3fs-package}
\alias{s3fs}
\alias{s3fs-package}
\title{s3fs: File System like `API` for Amazon Web Service `S3`.}
\description{
Access Amazon Web Service `S3` as if it were a file system. File system `API` design around R package `fs`
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

}
