% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{s3fs-package}
\alias{s3fs}
\alias{s3fs-package}
\title{s3fs: 'Amazon Web Service S3' File System}
\description{
Access 'Amazon Web Service Simple Storage Service' ('S3') \url{https://aws.amazon.com/s3/} as if it were a file system. Interface based on the R package 'fs'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DyfanJones/s3fs}
  \item Report bugs at \url{https://github.com/DyfanJones/s3fs/issues}
}

}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

}
